Add ability to pass the VAST response XML document in the options
In some cases, you may have the VAST response in hand, and you don't
want the VAST client to fetch it from an URL but still want to let it
resolve wrappers if any. This PR let you passe the XML VAST response
as the `response` field of the `options` parameter. In this case, the
`url` parameter is ignored and the provided response is returned to
the parser.